+++ client/vite.config.js
@@ -2,10 +2,20 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import tailwindcss from '@tailwindcss/vite';
 import svgr from 'vite-plugin-svgr';
+import mdx from '@mdx-js/rollup';
+import { visualizer } from 'rollup-plugin-visualizer';
 
 // https://vite.dev/config/
 export default defineConfig({
+  define: {
+    'process.env.NODE_ENV': '"production"',
+  },
+
   plugins: [
+    {
+      enforce: 'pre',
+      ...mdx(),
+    },
     react(),
     tailwindcss(),
     svgr({
@@ -15,7 +25,14 @@ export default defineConfig({
         },
       },
     }),
+
+    visualizer({
+      open: true, // automatically opens the report
+      gzipSize: true, // shows gzip sizes
+      brotliSize: true, // shows brotli sizes
+    }),
   ],
+
   server: {
     port: 3000,
   },